requirements.txt

jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pandas==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.45
sqlparse==0.5.4
